LUCENE-5207: Simpler without cast
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene5207@1523361 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/expressions/src/java/org/apache/lucene/expressions/js/JavascriptCompiler.java
@@ -531,7 +531,7 @@ public class JavascriptCompiler {
     if (expected == Type.INT_TYPE) {
         methodVisitor.push(truth);
     } else if (expected == Type.LONG_TYPE) {
-        methodVisitor.push((long) (truth ? 1 : 0));
+        methodVisitor.push(truth ? 1L : 0L);
     } else if (expected == Type.DOUBLE_TYPE) {
         methodVisitor.push(truth ? 1. : 0.);
     } else {
